Improve Integration test of Rest Controllers by decoupling queries from DB
- Originally, the GetByID method's test depended on a specific ID from the
database BUT there's a simpler solution, specifically to grab an ID from
the full list of DB entities. Assuming the GetByID method works correctly
then the given endpoint and ID should make the test pass. As a bonus, the
test should be less brittle since it is no longer coupled to the actual DB.